+++ actix-router/CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+- Minimum supported Rust version (MSRV) is now 1.56.
 
 
 ## 0.5.0 - 2022-02-22