+++ actix-router/CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
 ## Unreleased - 2022-xx-xx
+- Fix typo in error string in `Deserializer::deserialize_i32` implementation for `Value`. [#2876]
 - Minimum supported Rust version (MSRV) is now 1.59 due to transitive `time` dependency.
+
+[#2876]: https://github.com/actix/actix-web/pull/2876
 
 
 ## 0.5.0 - 2022-02-22